+++ src/components/menu-item/menu-item.styles.ts
@@ -65,10 +65,11 @@ export default css`
     color: var(--sl-color-neutral-1000);
   }
 
-  :host(:focus-visible:not([aria-disabled='true'])) .menu-item {
+  :host(:focus-visible) .menu-item {
     outline: none;
     background-color: var(--sl-color-primary-600);
     color: var(--sl-color-neutral-0);
+    opacity: 1;
   }
 
   .menu-item .menu-item__check,
@@ -88,7 +89,7 @@ export default css`
 
   @media (forced-colors: active) {
     :host(:hover:not([aria-disabled='true'])) .menu-item,
-    :host(:focus-visible:not([aria-disabled='true'])) .menu-item {
+    :host(:focus-visible) .menu-item {
       outline: dashed 1px SelectedItem;
       outline-offset: -1px;
     }